# openapi.yaml
openapi: 3.0.1
info:
  title: Sunsync
  description: 'Platform to monitoring the photovoltaic systems of the municipality of Trento'
  version: 1.0.0
...
paths:
...
  /api/v1/wsdata:
    get:
      summary: Get all weather station data
      deprecated: false
      description: >-
        Return all the weather station data. Can be filtered by date and weather
        station id.
      operationId: GetWSData
      tags:
        - WeatherStation/WsData
      parameters:
        - name: startdate
          in: query
          description: ''
          required: false
          schema:
            type: string
        - name: enddate
          in: query
          description: ''
          required: false
          schema:
            type: string
        - name: wsinfo_id
          in: query
          description: ''
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    ws_id:
                      type: integer
                    location:
                      type: object
                      properties:
                        lat:
                          type: number
                        long:
                          type: number
                        alt:
                          type: number
                      required:
                        - lat
                        - long
                        - alt
                    description:
                      type: string
                    url:
                      type: string
                  required:
                    - ws_id
                    - location
                    - description
                    - url
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        '404':
          description: Record Not Found
          content:
            application/json:
              schema:
                type: string
      security: []
...
  /api/v1/wsdata/{wsinfo_id}:
    get:
      summary: Get weather station data by Id
      deprecated: false
      description: Return weather station data by id.
      operationId: GetWSDataById
      tags:
        - WeatherStation/WsData
      parameters:
        - name: wsinfo_id
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    time:
                      type: integer
                    rain:
                      type: number
                    temperature:
                      type: number
                    humidity:
                      type: number
                    wind_direction:
                      type: number
                    wind_speed:
                      type: number
                    pressure:
                      type: number
                    solar_power:
                      type: number
                    ws_id:
                      type: integer
                  required:
                    - time
                    - rain
                    - temperature
                    - humidity
                    - wind_direction
                    - wind_speed
                    - pressure
                    - solar_power
                    - ws_id
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        '404':
          description: Record Not Found
          content:
            application/json:
              schema:
                type: string
      security: []
...
components:
...
  securitySchemes: {}